refactor the code to avoid copying the whole ical file to memory
This is how Evil_Bob does it with xml.c, a sane and tiny XML parser.

The document structure gets parsed, and a struct parser passed to the
parser contains function pointers. As parsed objects get encountered,
the matching function pointers gets called with the parsed text as
parameter. Then the content can be processed as it is read, instead of
parsed first and processed after.